delete of a role should now properly remove all mandates related to that role

+++ routes/roles.js
@@ -52,12 +52,19 @@ router.post('/update', helpers.requireadmin, function(req, res) {
 });
 
 router.post('/delete/:id', helpers.requireadmin, function(req, res) {
-  //TODO: this leaves old mandates left in the system. Good enough for now.
   var id = req.params.id;
-  models.Role.destroy({
-    where: {id: id, identifier: req.body.identifier} //Additional verification so less accidental deletes.
-  }).then(function() {
-    res.redirect('/');
+  models.Role.findOne({where: {id: id, identifier: req.body.identifier}}).then(function(role) {
+    if(role) {
+      //Correct identifier now do the delete.
+      //Destroy the role
+      models.Role.destroy({
+        where: {id: id}
+      });
+      res.redirect('/'); //Success
+    } else {
+      res.status(403);
+      res.send("failed verification");
+    }
   });
 });
